server/requirements.txt
asgiref==3.11.1
pymysql==1.1.2
Django==6.0.3
django-cors-headers==4.9.0
django-jazzmin==3.0.4
django-rest-framework==0.1.0
djangorestframework==3.17.0
pillow==12.1.1
sqlparse==0.5.5
tzdata==2025.3
firebase-admin==6.5.0
